+++ R/manystates-states.R
@@ -1,38 +1,22 @@
 #' States datacube
-#' @description The `states` datacube is a list containing 
-#'   `r length(manystates::states)` datasets: 
-#'   `r cli::pluralize("{names(manystates::states)}")`.
+#' @description `r describe_datacube(manystates::states)`.
 #'   It is a work-in-progress, so please let us know if you have any comments or suggestions.
 #' @format
 #' \describe{
+#' \item{ISD: }{A dataset with `r prettyNum(nrow(manystates::states$ISD), big.mark=",")` 
+#' observations and `r ncol(manystates::states$ISD)` variables: 
+#' `r cli::pluralize("{names(manystates::states$ISD)}")`.}
 #' \item{HUGGO: }{A dataset with `r prettyNum(nrow(manystates::states$HUGGO), big.mark=",")` 
 #' observations and `r ncol(manystates::states$HUGGO)` variables: 
 #' `r cli::pluralize("{names(manystates::states$HUGGO)}")`.}
 #' \item{GW: }{A dataset with `r prettyNum(nrow(manystates::states$GW), big.mark=",")` 
 #' observations and `r ncol(manystates::states$GW)` variables: 
 #' `r cli::pluralize("{names(manystates::states$GW)}")`.}
-#' \item{ISD: }{A dataset with `r prettyNum(nrow(manystates::states$ISD), big.mark=",")` 
-#' observations and `r ncol(manystates::states$ISD)` variables: 
-#' `r cli::pluralize("{names(manystates::states$ISD)}")`.}
 #' }
 #' For more information and references to each of the datasets used,
 #' please use the `manydata::call_sources()` and `manydata::compare_dimensions()` functions.
 #' @source
-#' \describe{
-#' \item{HUGGO: }{
-#' Hollway, James, Henrique Sposito, and Jael Tan. 2021. _manystates: States for manydata_.
-#' }
-#' \item{GW: }{
-#' Gleditsch, K.S., and M. D. Ward. “Interstate system membership: A revised list of the independent states since 1816”.
-#' _International Interactions_ 25.4 (1999), pp. 393-413.
-#' \url{http://ksgleditsch.com/data-4.html}
-#' }
-#' \item{ISD: }{
-#' Griffiths, R.D., and C. R. Butcher. “Introducing the international system (s) dataset (ISD), 1816-2011”.
-#' _International Interactions_ 39.5 (2013), pp. 748-768.
-#' \url{http://www.ryan-griffiths.com/data}
-#' }
-#' }
+#'   `r call_citations(states, output = "help")`
 #' @section Mapping:
 #' 
 #' |  *manystates*  | *GW*  | *ISD* | 
@@ -50,3 +34,12 @@
 #' lapply(states, manydata::mreport)
 #' ```
 "states"
+
+info_states <- tibble::tibble(Dataset = names(utils::data(states, package = "manystates")),
+                               Source = c("Griffiths, Ryan D., and Charles R. Butcher. 'Introducing the international system(s) dataset (ISD), 1816-2011'. International Interactions 39.5 (2013), pp. 748-768.",
+                                          "Hollway, James, Henrique Sposito, and Jael Tan. 2021. States for manydata.",
+                                          "Gleditsch, Kristian S., and Michael D. Ward. 'Interstate system membership: A revised list of the independent states since 1816'. International Interactions 25.4 (1999), pp. 393-413."),
+                               URL = c("http://www.ryan-griffiths.com/data",
+                                       "",
+                                       "http://ksgleditsch.com/data-4.html"))
+