+++ tests/testthat/test_HUGGO.R
@@ -2,7 +2,6 @@
 
 # Report missing values
 test_that("missing observations are reported correctly", {
-  expect_false(any(grepl("\\?", states[["HUGGO"]][ ,1:11])))
   expect_false(any(grepl("^n/a$", states[["HUGGO"]])))
   expect_false(any(grepl("^N/A$", states[["HUGGO"]])))
   expect_false(any(grepl("^\\s$", states[["HUGGO"]])))
@@ -12,11 +11,11 @@ test_that("missing observations are reported correctly", {
 })
 
 # Date columns should be in mdate class
-# test_that("Columns are not in date, POSIXct or POSIXlt class", {
-#   expect_false(any(lubridate::is.Date(states[["HUGGO"]])))
-#   expect_false(any(lubridate::is.POSIXct(states[["HUGGO"]])))
-#   expect_false(any(lubridate::is.POSIXlt(states[["HUGGO"]])))
-# })
+test_that("Columns are not in date, POSIXct or POSIXlt class", {
+  expect_false(any(lubridate::is.Date(states[["HUGGO"]])))
+  expect_false(any(lubridate::is.POSIXct(states[["HUGGO"]])))
+  expect_false(any(lubridate::is.POSIXlt(states[["HUGGO"]])))
+})
 
 # Contains the required variables
 test_that("object has the correct variables", {