+++ tests/testthat/test_ISD.R
@@ -2,7 +2,6 @@
 
 # Report missing values
 test_that("missing observations are reported correctly", {
-  expect_false(any(grepl("\\?", states[["ISD"]])))
   expect_false(any(grepl("^n/a$", states[["ISD"]])))
   expect_false(any(grepl("^N/A$", states[["ISD"]])))
   expect_false(any(grepl("^\\s$", states[["ISD"]])))
@@ -12,11 +11,11 @@ test_that("missing observations are reported correctly", {
 })
 
 # Date columns should be in mdate class
-# test_that("Columns are not in date, POSIXct or POSIXlt class", {
-#   expect_false(any(lubridate::is.Date(states[["ISD"]])))
-#   expect_false(any(lubridate::is.POSIXct(states[["ISD"]])))
-#   expect_false(any(lubridate::is.POSIXlt(states[["ISD"]])))
-# })
+test_that("Columns are not in date, POSIXct or POSIXlt class", {
+  expect_false(any(lubridate::is.Date(states[["ISD"]])))
+  expect_false(any(lubridate::is.POSIXct(states[["ISD"]])))
+  expect_false(any(lubridate::is.POSIXlt(states[["ISD"]])))
+})
 
 # Contains the required variables
 test_that("object has the correct variables", {
@@ -38,7 +37,7 @@ test_that("dates are standardised", {
   expect_false(any(grepl("/", states[["ISD"]]$End)))
 })
 
-# Labels are standardized
+# State names are standardized
 test_that("labels are standardised", {
   expect_false(any(grepl("U\\.S\\.", states[["ISD"]])))
   expect_false(any(grepl("U\\.K\\.", states[["ISD"]])))