+++ NEWS.md
@@ -1,3 +1,17 @@
+# manystates 0.3.1
+
+## Package
+
+- Moved `code_states()` function from `{manypkgs}` to be user-facing
+- Added `code_states()` to website
+- Improved regex matches for `code_states()`
+- Added tests for `code_states()`
+
+## Data
+
+- Closed #72 by adding version 2 of `states$ISD`
+- Closed #64 by filling in missing state names in `states$HUGGO`
+
 # manystates 0.3.0
 
 ## Package